+++ NEWS.md
@@ -6,6 +6,10 @@ Breaking change:
 
 * `theme_void()` calls `strip_tt()` to drop every change made to this point, so the order of operations matter.
 
+New:
+
+* A global option can be used to choose the temporary directory where HTML files are saved for preview: `options(tinytable_tempdir = "/home/username/tempdir")`
+
 ## 0.13.0
 
 * New aliases to facilitate completion in IDEs: `tt_style()`, `tt_format()`, `tt_plot()`, `tt_group()`, `tt_save()`. Thanks to @rpruim for the suggestion in Issue #540.